+++ nix/flake-module.nix
@@ -63,6 +63,20 @@ in
                   Build tools useful for Haskell development are included by default.
                 '';
               };
+              extraLibraries = mkOption {
+                type = functionTo (types.attrsOf (types.nullOr types.package));
+                description = ''
+                  Extra Haskell libraries available in the shell's environment.
+                  These can be used in the shell's `runghc` and `ghci` for instance.
+
+                  The argument is the Haskell package set.
+
+                  The return type is an attribute set for overridability and syntax, as only the values are used.
+                '';
+                default = hp: { };
+                defaultText = lib.literalExpression "hp: { }";
+                example = lib.literalExpression "hp: { inherit (hp) releaser; }";
+              };
               hlsCheck = mkOption {
                 default = { };
                 type = hlsCheckSubmodule;